+++ resolwe/flow/models/entity.py
@@ -377,6 +377,15 @@ class Entity(BaseCollection, PermissionObject):
         if validation_errors:
             raise ValidationError(validation_errors)
 
+    def save(self, *args, **kwargs):
+        """Propagate the modified time to the collection."""
+        super().save(*args, **kwargs)
+        if self.collection:
+            self.collection.skip_auto_now = True  # type: ignore
+            self.collection.modified = self.modified
+            self.collection.save(update_fields=["modified"])
+            self.collection.skip_auto_now = False  # type: ignore
+
 
 class RelationType(models.Model):
     """Model for storing relation types."""
@@ -501,6 +510,11 @@ class Relation(BaseModel, PermissionObject):
                 "`descriptor_schema` must be defined if `descriptor` is given"
             )
         super().save()
+        if self.collection:
+            self.collection.skip_auto_now = True  # type: ignore
+            self.collection.modified = self.modified
+            self.collection.save(update_fields=["modified"])
+            self.collection.skip_auto_now = False  # type: ignore
 
 
 class RelationPartition(models.Model):